+++ src/components/org-users/org-users.test.ts
@@ -66,38 +66,22 @@ describe('org-users test suite', () => {
   let nockAccounts: nock.Scope;
 
   beforeEach(() => {
+    nock.cleanAll();
+
     nockAccounts = nock(ctx.app.accountsAPI);
     nockCF = nock(ctx.app.cloudFoundryAPI);
     nockUAA = nock(ctx.app.uaaAPI);
-
     nockNotify = nock(/api.notifications.service.gov.uk/)
       .filteringPath(() => '/');
-
-    // tslint:disable:max-line-length
-    nockCF
-      .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/auditors/uaa-user-edit-123456')
-      .reply(200, `{}`)
-    ;
-    // tslint:enable:max-line-length
-  });
-
-  afterAll(() => {
-    nockAccounts.done();
-    nockCF.done();
-    nockNotify.done();
-    nockUAA.done();
   });
 
   afterEach(() => {
-    expect(nockAccounts.isDone()).toBeTruthy();
-    expect(nockNotify.isDone()).toBeTruthy();
-    expect(nockUAA.isDone()).toBeTruthy();
-    //expect(nockCF.isDone()).toBeTruthy();
-
     nockAccounts.done();
     nockNotify.done();
     nockUAA.done();
-    //nockCF.done();
+    nockCF.done();
+
+    nock.cleanAll();
   });
 
   it('should show the users pages', async () => {
@@ -133,8 +117,6 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg);
 
     const response = await orgUsers.inviteUserForm(ctx, {
@@ -154,8 +136,7 @@ describe('org-users test suite', () => {
       .reply(200, cfData.organization)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+    ;
 
     const response = await orgUsers.inviteUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -175,8 +156,7 @@ describe('org-users test suite', () => {
       .reply(200, cfData.organization)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+    ;
 
     const response = await orgUsers.inviteUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -234,18 +214,6 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg);
 
-    nockCF
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
-      .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
-
     const response = await orgUsers.inviteUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
     }, {email: 'jeff@jeff.com'});
@@ -265,8 +233,6 @@ describe('org-users test suite', () => {
       .reply(200, cfData.organization)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/users')
       .reply(201, `{"metadata": {"guid": "3deb9f04-b449-4f94-b3dd-c73cefe5b275"}}`)
-      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/billing_managers/uaa-user-edit-123456?recursive=true')
-      .reply(200, `{}`)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/billing_managers/5ff19d4c-8fa0-4d74-94e0-52eac86d55a8?recursive=true')
       .reply(200, `{}`)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
@@ -318,8 +284,6 @@ describe('org-users test suite', () => {
       .reply(200, cfData.organization)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/users')
       .reply(201, `{"metadata": {"guid": "3deb9f04-b449-4f94-b3dd-c73cefe5b275"}}`)
-      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/managers/uaa-user-edit-123456?recursive=true')
-      .reply(200, `{}`)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/managers/5ff19d4c-8fa0-4d74-94e0-52eac86d55a8?recursive=true')
       .reply(200, `{}`)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
@@ -327,7 +291,8 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+      .reply(200, cfData.userRolesForOrg)
+    ;
 
     nockUAA
       .post('/invite_users?redirect_uri=https://www.cloud.service.gov.uk/next-steps?success&client_id=user_invitation')
@@ -371,8 +336,6 @@ describe('org-users test suite', () => {
       .reply(200, cfData.organization)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/users')
       .reply(201, `{"metadata": {"guid": "3deb9f04-b449-4f94-b3dd-c73cefe5b275"}}`)
-      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/auditors/uaa-user-edit-123456?recursive=true')
-      .reply(200, `{}`)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/auditors/5ff19d4c-8fa0-4d74-94e0-52eac86d55a8?recursive=true')
       .reply(200, `{}`)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
@@ -474,12 +437,8 @@ describe('org-users test suite', () => {
       .reply(200, cfData.organization)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/users')
       .reply(201, `{"metadata": {"guid": "3deb9f04-b449-4f94-b3dd-c73cefe5b275"}}`)
-      .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/managers/uaa-user-edit-123456')
-      .reply(200, `{}`)
       .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/developers/5ff19d4c-8fa0-4d74-94e0-52eac86d55a8')
       .reply(200, `{}`)
-      .post('/oauth/token?grant_type=client_credentials')
-      .reply(200, `{"access_token": "FAKE_ACCESS_TOKEN"}`)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
       .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
@@ -532,8 +491,6 @@ describe('org-users test suite', () => {
       .reply(200, `{}`)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
       .reply(200, cfData.spaces)
-      .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/developers/uaa-user-edit-123456')
-      .reply(200, `{}`)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
@@ -581,8 +538,6 @@ describe('org-users test suite', () => {
       .reply(200, cfData.organization)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/users')
       .reply(201, `{"metadata": {"guid": "3deb9f04-b449-4f94-b3dd-c73cefe5b275"}}`)
-      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/users/5ff19d4c-8fa0-4d74-94e0-52eac86d55a8?recursive=true')
-      .reply(200, `{}`)
       .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/billing_managers/5ff19d4c-8fa0-4d74-94e0-52eac86d55a8?recursive=true')
       .reply(200, `{}`)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
@@ -624,10 +579,6 @@ describe('org-users test suite', () => {
 
   it('should fail if the user does not exist in org', async () => {
     nockCF
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
@@ -656,14 +607,11 @@ describe('org-users test suite', () => {
     nockCF
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+    ;
 
     const response = await orgUsers.resendInvitation(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -677,8 +625,6 @@ describe('org-users test suite', () => {
     nockCF
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
@@ -697,20 +643,15 @@ describe('org-users test suite', () => {
   it('should update the user, set BillingManager role and show success - User Edit', async () => {
     // tslint:disable:max-line-length
     nockCF
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
       .delete('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/users/5ff19d4c-8fa0-4d74-94e0-52eac86d55a8?recursive=true')
       .reply(200, {})
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+      .reply(200, cfData.userRolesForOrg)
+    ;
     // tslint:enable:max-line-length
 
     const response = await orgUsers.deleteUser(ctx, {
@@ -723,9 +664,11 @@ describe('org-users test suite', () => {
 
   it('should show the user edit page', async () => {
     nockUAA
-      .get('/Users/uaa-user-edit-123456').reply(200, uaaData.usersByEmail)
+      .get('/Users/uaa-user-edit-123456')
+      .reply(200, uaaData.usersByEmail)
       .post('/oauth/token?grant_type=client_credentials')
-      .reply(200, `{"access_token": "FAKE_ACCESS_TOKEN"}`);
+      .reply(200, `{"access_token": "FAKE_ACCESS_TOKEN"}`)
+    ;
 
     nockAccounts.get('/users/uaa-user-edit-123456').reply(200, `{
       "user_uuid": "uaa-user-edit-123456",
@@ -747,7 +690,10 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+      .reply(200, cfData.userRolesForOrg)
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
+      .reply(200, cfData.userRolesForOrg)
+    ;
 
     const response = await orgUsers.editUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -769,7 +715,10 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+      .reply(200, cfData.userRolesForOrg)
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
+      .reply(200, cfData.userRolesForOrg)
+    ;
 
     await expect(orgUsers.editUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -803,16 +752,19 @@ describe('org-users test suite', () => {
     nockCF
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+      .reply(200, cfData.userRolesForOrg)
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
+      .reply(200, cfData.spaces)
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
+      .reply(200, cfData.spaces)
+      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/billing_managers/uaa-user-edit-123456?recursive=true')
+      .reply(200, `{}`)
+    ;
 
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -847,13 +799,11 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
-
-    const scope = nock(ctx.app.cloudFoundryAPI)
-      // tslint:disable:max-line-length
-        .delete('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/billing_managers/uaa-user-changeperms-123456?recursive=true').reply(200, `{}`)
-      // tslint:enable:max-line-length
+      .reply(200, cfData.userRolesForOrg)
+      .delete('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/billing_managers/uaa-user-changeperms-123456?recursive=true')
+      .reply(200, `{}`)
     ;
+
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
       userGUID: 'uaa-user-changeperms-123456',
@@ -870,8 +820,6 @@ describe('org-users test suite', () => {
       },
     });
     expect(response.body).toContain('Updated a team member');
-    expect(scope.isDone());
-    scope.done();
   });
 
   it('should update the user, set OrgManager role and show success - User Edit', async () => {
@@ -887,7 +835,10 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+      .reply(200, cfData.userRolesForOrg)
+      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/managers/uaa-user-edit-123456?recursive=true')
+      .reply(200, `{}`)
+    ;
 
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -922,13 +873,11 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
-
-    const scope = nock(ctx.app.cloudFoundryAPI)
-      // tslint:disable:max-line-length
-        .delete('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/managers/uaa-user-changeperms-123456?recursive=true').reply(200, `{}`)
-      // tslint:enable:max-line-length
+      .reply(200, cfData.userRolesForOrg)
+      .delete('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/managers/uaa-user-changeperms-123456?recursive=true')
+      .reply(200, `{}`)
     ;
+
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
       userGUID: 'uaa-user-changeperms-123456',
@@ -949,8 +898,6 @@ describe('org-users test suite', () => {
       },
     });
     expect(response.body).toContain('Updated a team member');
-    expect(scope.isDone());
-    scope.done();
   });
 
   it('should update the user, set OrgAuditor role and show success - User Edit', async () => {
@@ -966,7 +913,10 @@ describe('org-users test suite', () => {
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
       .reply(200, cfData.userRolesForOrg)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+      .reply(200, cfData.userRolesForOrg)
+      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/auditors/uaa-user-edit-123456?recursive=true')
+      .reply(200, `{}`)
+    ;
 
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -990,24 +940,21 @@ describe('org-users test suite', () => {
 
   it('should update the user, remove OrgAuditor role and show success - User Edit', async () => {
     nockCF
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
 
-    const scope = nock(ctx.app.cloudFoundryAPI)
-      // tslint:disable:max-line-length
-        .delete('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/auditors/uaa-user-changeperms-123456?recursive=true').reply(200, `{}`)
-      // tslint:enable:max-line-length
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
+      .times(3)
+      .reply(200, cfData.userRolesForOrg)
+
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
+      .times(2)
+      .reply(200, cfData.spaces)
+
+      .delete('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/auditors/uaa-user-changeperms-123456?recursive=true')
+      .reply(200, `{}`)
     ;
+
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
       userGUID: 'uaa-user-changeperms-123456',
@@ -1023,25 +970,26 @@ describe('org-users test suite', () => {
         '5489e195-c42b-4e61-bf30-323c331ecc01': composeSpaceRoles({}),
       },
     });
+
     expect(response.body).toContain('Updated a team member');
-    expect(scope.isDone());
-    scope.done();
   });
 
   it('should update the user, set SpaceManager role and show success - User Edit', async () => {
     nockCF
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
+
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
+      .times(3)
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
+      .times(2)
+      .reply(200, cfData.spaces)
+
+      .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/managers/uaa-user-edit-123456')
+      .reply(200, `{}`)
+    ;
 
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -1065,18 +1013,20 @@ describe('org-users test suite', () => {
 
   it('should update the user, set SpaceDeveloper role and show success - User Edit', async () => {
     nockCF
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
+
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
+      .times(3)
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
+      .times(2)
+      .reply(200, cfData.spaces)
+
+      .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/developers/uaa-user-edit-123456')
+      .reply(200, `{}`)
+    ;
 
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -1100,18 +1050,20 @@ describe('org-users test suite', () => {
 
   it('should update the user, set SpaceAuditor role and show success - User Edit', async () => {
     nockCF
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
       .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
+
       .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
+      .times(3)
       .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg);
+
+      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
+      .times(2)
+      .reply(200, cfData.spaces)
+
+      .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/auditors/uaa-user-edit-123456')
+      .reply(200, `{}`)
+    ;
 
     const response = await orgUsers.updateUser(ctx, {
       organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
@@ -1131,316 +1083,5 @@ describe('org-users test suite', () => {
     });
 
     expect(response.body).toContain('Updated a team member');
-  });
-});
-
-describe('permissions calling cc api', () => {
-  beforeEach(() => {
-    nock.cleanAll();
-  });
-
-  it('should make a single request due to permission update', async () => {
-    const scope = nock(ctx.app.cloudFoundryAPI).persist()
-      .put('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/managers/uaa-user-edit-123456?recursive=true')
-      .reply(200, `{}`)
-      .put('/v2/spaces/5489e195-c42b-4e61-bf30-323c331ecc01/developers/uaa-user-edit-123456')
-      .reply(200, `{}`)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
-      .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
-    ;
-
-    const response = await orgUsers.updateUser(ctx, {
-      organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
-      userGUID: 'uaa-user-edit-123456',
-    }, {
-      org_roles: {
-        '3deb9f04-b449-4f94-b3dd-c73cefe5b275': composeOrgRoles({
-          managers: {
-            current: '0',
-            desired: '1',
-          },
-        }),
-      },
-      space_roles: {
-        '5489e195-c42b-4e61-bf30-323c331ecc01': composeSpaceRoles({
-          developers: {
-            current: '0',
-            desired: '1',
-          },
-        }),
-      },
-    });
-
-    expect(scope.isDone()).toBeTruthy();
-    expect(response.body).toContain('Updated a team member');
-  });
-
-  it('should make no requests when permission has been previously and still is set', async () => {
-    const scope = nock('https://example.com/api').persist()
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
-      .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
-    ;
-
-    const response = await orgUsers.updateUser(ctx, {
-      organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
-      userGUID: 'uaa-user-edit-123456',
-    }, {
-      org_roles: {
-        '3deb9f04-b449-4f94-b3dd-c73cefe5b275': composeOrgRoles({
-          managers: {
-            current: '1',
-            desired: '1',
-          },
-        }),
-      },
-      space_roles: {
-        '5489e195-c42b-4e61-bf30-323c331ecc01': composeSpaceRoles({
-          developers: {
-            current: '1',
-            desired: '1',
-          },
-        }),
-      },
-    });
-
-    expect(scope.isDone()).toBeTruthy();
-    expect(response.body).toContain('Updated a team member');
-  });
-
-  it('should make no requests when permission has been previously and still is unset', async () => {
-    const scope = nock('https://example.com/api').persist()
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/user_roles')
-      .reply(200, cfData.userRolesForOrg)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275')
-      .reply(200, cfData.organization)
-      .get('/v2/organizations/3deb9f04-b449-4f94-b3dd-c73cefe5b275/spaces')
-      .reply(200, cfData.spaces)
-    ;
-
-    const response = await orgUsers.updateUser(ctx, {
-      organizationGUID: '3deb9f04-b449-4f94-b3dd-c73cefe5b275',
-      userGUID: 'uaa-user-edit-123456',
-    }, {
-      org_roles: {
-        '3deb9f04-b449-4f94-b3dd-c73cefe5b275': composeOrgRoles({
-          billing_managers: {
-            current: '0',
-          },
-        }),
-      },
-      space_roles: {
-        '5489e195-c42b-4e61-bf30-323c331ecc01': composeSpaceRoles({
-          developers: {
-            current: '0',
-          },
-        }),
-      },
-    });
-
-    expect(scope.isDone()).toBeTruthy();
-    expect(response.body).toContain('Updated a team member');
-  });
-});
-
-describe('_getUserRolesByGuid', () => {
-  let nockAccounts: nock.Scope;
-
-  beforeEach(() => {
-    nockAccounts = nock(ctx.app.accountsAPI).persist();
-
-    nockAccounts
-      .get('/users/some-user-guid-0').reply(200, JSON.stringify({
-      user_uuid: 'some-user-guid-0',
-      user_email: 'some-user-guid-0@fake.digital.cabinet-office.gov.uk',
-      username: 'some-fake-username-from-paas-accounts',
-    }))
-      .get('/users/some-user-guid-1').reply(404)
-      .get('/users/some-user-guid-2').reply(404);
-  });
-
-  it('should return an empty map if there are no users', async () => {
-    const userOrgRoles: any = [];
-    const spaceUserLists: any = [];
-
-    const accountsClient = new AccountsClient({
-      apiEndpoint: ctx.app.accountsAPI,
-      secret: ctx.app.accountsSecret,
-      logger: ctx.app.logger,
-    });
-
-    const result = await orgUsers._getUserRolesByGuid(userOrgRoles, spaceUserLists, accountsClient);
-    expect(result).toEqual({});
-  });
-
-  it('should return org roles of a user that has no space access', async () => {
-    const userOrgRoles: any = [
-      {
-        metadata: {guid: 'some-user-guid'},
-        entity: {organization_roles: ['org_manager'], username: 'some-user-name'},
-      },
-    ];
-    const spaceUserLists: any = [];
-
-    const accountsClient = new AccountsClient({
-      apiEndpoint: ctx.app.accountsAPI,
-      secret: ctx.app.accountsSecret,
-      logger: ctx.app.logger,
-    });
-
-    const result = await orgUsers._getUserRolesByGuid(userOrgRoles, spaceUserLists, accountsClient);
-    expect(result).toEqual({
-      'some-user-guid': {
-        orgRoles: ['org_manager'],
-        username: 'some-user-name',
-        spaces: [],
-      },
-    });
-  });
-
-  it('should return roles and space of a user that has access to one space', async () => {
-    const userOrgRoles: any = [
-      {
-        metadata: {guid: 'some-user-guid'},
-        entity: {organization_roles: ['org_manager'], username: 'some-user-name'},
-      },
-    ];
-
-    const space = {metadata: {guid: 'some-space-guid'}} as any;
-    const user = {metadata: {guid: 'some-user-guid'}} as any;
-
-    const spaceUserLists = [{
-      space,
-      users: [user],
-    }];
-
-    const accountsClient = new AccountsClient({
-      apiEndpoint: ctx.app.accountsAPI,
-      secret: ctx.app.accountsSecret,
-      logger: ctx.app.logger,
-    });
-
-    const result = await orgUsers._getUserRolesByGuid(userOrgRoles, spaceUserLists, accountsClient);
-    expect(result).toEqual({
-      'some-user-guid': {
-        orgRoles: ['org_manager'],
-        username: 'some-user-name',
-        spaces: [space],
-      },
-    });
-  });
-
-  it('should return roles and spaces of a user that has access to multiple spaces', async () => {
-    const userOrgRoles: any = [
-      {
-        metadata: {guid: 'some-user-guid'},
-        entity: {organization_roles: ['org_manager'], username: 'some-user-name'},
-      },
-    ];
-
-    const spaces = [1, 2, 3].map(i => ({metadata: {guid: `some-space-guid-${i}`}})) as any[];
-    const user = {metadata: {guid: 'some-user-guid'}} as any;
-
-    const spaceUserLists = spaces.map(space => ({
-      space,
-      users: [user],
-    }));
-
-    const accountsClient = new AccountsClient({
-      apiEndpoint: ctx.app.accountsAPI,
-      secret: ctx.app.accountsSecret,
-      logger: ctx.app.logger,
-    });
-
-    const result = await orgUsers._getUserRolesByGuid(userOrgRoles, spaceUserLists, accountsClient);
-    expect(result).toEqual({
-      'some-user-guid': {
-        orgRoles: ['org_manager'],
-        username: 'some-user-name',
-        spaces,
-      },
-    });
-  });
-
-  it('should return users, roles and spaces of multiple users', async () => {
-    const userOrgRoles: any = [0, 1, 2].map(i => (
-      {
-        metadata: {guid: `some-user-guid-${i}`},
-        entity: {organization_roles: ['org_manager'], username: `some-user-name-${i}`},
-      }
-    ));
-
-    const space: any = (i: number) => ({metadata: {guid: `some-space-guid-${i}`}});
-    const user: any = (i: number) => ({metadata: {guid: `some-user-guid-${i}`}});
-
-    const spaceUserLists = [
-      {space: space(0), users: [user(0), user(1)]},
-      {space: space(1), users: [user(1), user(2)]},
-      {space: space(2), users: [user(0), user(1), user(2)]},
-    ];
-
-    const accountsClient = new AccountsClient({
-      apiEndpoint: ctx.app.accountsAPI,
-      secret: ctx.app.accountsSecret,
-      logger: ctx.app.logger,
-    });
-
-    const result = await orgUsers._getUserRolesByGuid(userOrgRoles, spaceUserLists, accountsClient);
-    expect(result).toEqual({
-      'some-user-guid-0': {
-        orgRoles: ['org_manager'],
-        username: 'some-fake-username-from-paas-accounts',
-        spaces: [space(0), space(2)],
-      },
-      'some-user-guid-1': {
-        orgRoles: ['org_manager'],
-        username: 'some-user-name-1',
-        spaces: [space(0), space(1), space(2)],
-      },
-      'some-user-guid-2': {
-        orgRoles: ['org_manager'],
-        username: 'some-user-name-2',
-        spaces: [space(1), space(2)],
-      },
-    });
-  });
-
-  it('should get the user\'s username from accounts, falling back to UAA', async () => {
-    const userOrgRoles: any = [0, 1].map(i => (
-      {
-        metadata: {guid: `some-user-guid-${i}`},
-        entity: {organization_roles: ['org_manager'], username: `some-user-name-${i}`},
-      }
-    ));
-
-    const space: any = (i: number) => ({metadata: {guid: `some-space-guid-${i}`}});
-    const user: any = (i: number) => ({metadata: {guid: `some-user-guid-${i}`}});
-
-    const spaceUserLists = [
-      {space: space(0), users: [user(0), user(1)]},
-    ];
-
-    const accountsClient = new AccountsClient({
-      apiEndpoint: ctx.app.accountsAPI,
-      secret: ctx.app.accountsSecret,
-      logger: ctx.app.logger,
-    });
-
-    const result = await orgUsers._getUserRolesByGuid(userOrgRoles, spaceUserLists, accountsClient);
-    expect(result).toMatchObject({
-      'some-user-guid-0': {
-        username: 'some-fake-username-from-paas-accounts',
-      },
-      'some-user-guid-1': {
-        username: 'some-user-name-1',
-      },
-    });
   });
 });
